requirements.txt
Django==5.0.6
Gunicorn==22.0.0
mediapipe==0.10.14
numpy==1.26.4
opencv_contrib_python==4.9.0.80
tensorflow==2.14.0
python-dotenv==1.0.1
